+++ tests/Webfactory/Doctrine/ORMTestInfrastructure/_files/ORMInfrastructure/Annotation/CustomAnnotation.php
@@ -1,5 +1,12 @@
 <?php
 
+/*
+ * (c) webfactory GmbH <info@webfactory.de>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Webfactory\Doctrine\ORMTestInfrastructure\ORMInfrastructureTest\Annotation;
 
 /**
