+++ tests/Webfactory/Doctrine/ORMTestInfrastructure/_files/ORMInfrastructure/Annotation/TestAnnotation.php
@@ -7,7 +7,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Webfactory\Doctrine\ORMTestInfrastructure\ORMInfrastructureTest;
+namespace Webfactory\Doctrine\ORMTestInfrastructure\ORMInfrastructureTest\Annotation;
 
 /**
  * An annotation that is used in the test scenarios.
