+++ tests/Webfactory/Doctrine/ORMTestInfrastructure/_files/ORMInfrastructure/Entity/AnnotatedTestEntity.php
@@ -10,6 +10,7 @@
 namespace Webfactory\Doctrine\ORMTestInfrastructure\ORMInfrastructureTest;
 
 use Doctrine\ORM\Mapping as ORM;
+use \Webfactory\Doctrine\ORMTestInfrastructure\ORMInfrastructureTest\Annotation\TestAnnotation;
 
 /**
  * An entity that uses a custom annotation.
